[Squashed] add Battery Styles the new way 1/2
google draws now the battery icon which fits perfectly into our
concept. Beside that the batteryview is as well used on QS.
So we readd our battery icons (icon, percentage, icon with percentage
circle, circle perc, dot circle, dot circle perc) back and all are
drawn. As well the icon does not only change on statusbar as well
on QS it is reflected.

As usual color options are present for the User to style the icon

Change-Id: Iae9021af56a21b8d39efc5162186eb17ecc933e2

Conflicts:

	core/java/android/provider/Settings.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java

Conflicts:
	core/java/android/provider/Settings.java
	packages/SystemUI/res/layout/status_bar.xml
	packages/SystemUI/res/values/mahdi_strings.xml
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettings.java

Commit 1: fb: show percentage on icon in bold

and use for only percent more width in statusbar and QS
As well show on percent with icon on fully charge not the bolt
and instead the value 100

also add SettingsObserver to PhoneStatusBar.java

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java

Commit 2: fb: follow new 4.4 transition on center clock and circle battery

Change-Id: I50657071f290e0646193d079ab9319ed15f667b6

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBarTransitions.java

Commit 3: fb: fix rounding issues and take density into account

we are a multiDpi rom...and on 4.2 and 4.3 already the text
in the circle battery icon was on lower dpi a small bit up. The original
author compensated this with just a +1 to the Y value. Thats not the right
way...calculate the current dpi into the offset and all is well.

We want a perfect 4.4...so this nitpicks even if we are talking about
0.5dp maybe should be fixed

Change-Id: I835aa20b82aea980eb79a4772adccf30b2d6f79f

Commit 4: Fix battery icon with percentage showing "!" and percentage only color when under 14%

Change-Id: I559fd4d9fbb7f9599fe1b3cf75602b7ad7e2e301

Commit 5: fb: add back small bottom margin 1px

2px was to much 0 was to less.....1 seems to be perfect

Change-Id: Iefa3aee39197786da1652c3b372aa46c0ea3a6dc

Commit 6: Circle Battery: reduce the size a little

fb: add bottom margin to stock default battery icon

We set the layout parameters programmatically so the fix from
google does not take effect. To solve set the bottom margin
programmatically based on the current display density

Change-Id: I9caa0e729afa19eb7cb852a1df6ad036a2839d1b