Navigation Bar: Rotate the search light in landscape
The current search light icon is an arrow pointing upwards, so while in portrait it points to the direction to be swiped to access Google Now (and the customizable shortcuts), in landscape it points upward as well which looks out of place, and swiping upwards reveals that you have to swipe towards the left to access those targets. Rotate it by 270 degrees to make it point towards the right direction (i.e. Towards the lockscreen widgets)