[Squashed]On-The-Go mode [1/2]
On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

allows to change transparency of the overlay via the notification.

Fixes:
-on the go :fixes
-Base: On-The-Go: Allow to toggle between front and back camera
-Base: On-The-Go: Several fixes
-Base: On-The-Go: Overlay status- and navigation bar as well
-Base: On-The-Go: Change the way of changing the camera mode